requirements.txt
# App
gunicorn
motor
requests
appdirs==1.4.4
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
h2==3.2.0
hpack==3.0.0
httpcore==0.12.3
httpx==0.16.1
hyperframe==5.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
pathspec==0.8.1
urllib3==1.26.3
watchgod==0.6
python-multipart
Pillow
pybase64
anyio==3.6.2
click==8.1.3
fastapi==0.92.0
h11==0.14.0
idna==3.4
pydantic==1.10.5
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
# Mains
python-pptx==0.6.22
Spire.Presentation==8.9.0
Spire==0.4.2

python-pptx
fpdf
